Minor fixes - adapting build data set
